next basic rules
