+++ huskey-config-backup.txt
@@ -1,0 +1,11 @@
+,
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "./src/**/*.{js,jsx}": [
+      "eslint ./src --fix"
+    ]
+  }
