+++ src/pages/useMouse.js
@@ -1,4 +1,6 @@
 import { useMouse } from 'react-use'
+// import Link from 'next/link'
+import Router from 'next/router'
 
 const UseMouse = () => {
   const ref = React.useRef(null)
@@ -8,18 +10,21 @@ const UseMouse = () => {
   console.log(a)
 
   return (
-    <div ref={ref}>
-      <div>
-        Mouse position in document - x:{docX} y:{docY}
-      </div>
-      <div>
-        Mouse position in element - x:{elX} y:{elY}
-      </div>
-      <div>
-        Element position- x:{posX} y:{posY}
-      </div>
-      <div>
-        Element dimensions - {elW}x{elH}
+    <div>
+      <button onClick={() => Router.push('/index')}>Back to index</button>
+      <div ref={ref}>
+        <div>
+          Mouse position in document - x:{docX} y:{docY}
+        </div>
+        <div>
+          Mouse position in element - x:{elX} y:{elY}
+        </div>
+        <div>
+          Element position- x:{posX} y:{posY}
+        </div>
+        <div>
+          Element dimensions - {elW}x{elH}
+        </div>
       </div>
     </div>
   )
